Factor common code
